# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OsteoJP API
  version: "1.0.0"
  summary: Patient portal API + AI-partner ingestion contract for the OsteoJP platform.
  description: >
    Canonical contract for the OsteoJP V1 API. Two audiences share this document:

      * **Patient portal / external integrators** — the `api.osteojp.pt` patient
        API. Every patient route is self-scoped: the caller's `patient_id` and
        `tenant_id` are derived server-side from the verified token, never from
        request payload, and RLS re-enforces self-scope as defense in depth.
      * **AI ingestion partner** — the server-to-server
        `POST /api/v1/ingestion/clinical-records` endpoint, authenticated by an
        HMAC signature over the raw body (timestamp + signature header pair).

    This spec DOCUMENTS existing behavior; it does not define new endpoints.
    It is hand-authored (the handlers validate with hand-rolled guards, not
    importable zod schemas) and kept honest by a CI drift check that fails if any
    `/api/v1` route lacks a path here.

    Notes that hold across the patient API:
      * All timestamps are ISO 8601 UTC strings.
      * Money is integer cents with an explicit currency; never a float.
      * `force-dynamic`: responses are per-session and never cached.
  license:
    name: Proprietary
    url: https://osteojp.pt
  contact:
    name: OsteoJP Platform
    url: https://osteojp.pt
servers:
  - url: https://api.osteojp.pt
...
  - url: https://osteojp.pt
...
security:
  - patientBearer: []
paths:
...
  /api/v1/booking/guest:
    post:
      operationId: createGuestBookingRequest
      tags: [booking]
      summary: Request an appointment as a NEW client, with no account
      # CORRECTED 2026-08-14. This override was MISSING, so the document's
      # default (`patientBearer`) applied and the spec claimed the project's one
      # unauthenticated write surface required a patient JWT - the exact
      # opposite of what it is, in the document a reviewer reads to find out.
      # The route was always correct; the description of it was not.
      security: []
      description: >
        ITEM 6. A booking REQUEST from somebody who is not a patient: no
        account, no OTP login. The project's FIRST unauthenticated write
        surface.


        EVERY SUBMISSION IS A REQUEST (R-GUEST-1). Nothing is booked and no slot
        is held, whatever the availability. `status` is defaulted by the
        database and is never set by this route, so the path cannot be made to
        auto-confirm by passing a field. Reception confirms each one by hand and
        creates the patient and the appointment at that moment.


        NOTHING CLINICAL IS TOUCHED. The write goes to `guest_booking_requests`
        and nowhere else — no patient row, no appointment, no notification.
        That table has no anon policy in either direction; the insert runs
        service-role with an explicit `tenant_id`.


        IT ANSWERS 202 WHETHER OR NOT THE PHONE MATCHES AN EXISTING PATIENT, and
        never queries the patient table on this path at all, so membership
        cannot leak even through the timing of a lookup. The
        possible-existing-patient flag is computed for RECEPTION when the queue
        is read, and is never returned here. A response that differed would make
        a public form a patient-list oracle.


        NO SMS IS SENT (R9). The caller shows an on-screen confirmation and
        tells the person the clinic will make contact.


        IT TAKES A PREFERENCE, NOT A SLOT (Option A, 2026-08-14). `startsAt`,
        `endsAt` and `practitionerId` are REFUSED; `preferredDate` and
        `preferredPeriod` are required. See `GuestBookingInput` for why: with the
        roster and the slot list unexposed, an exact time arriving here could not
        have come from anything public.


        Rate limited on THREE independent axes — per source, per phone hash, and
        two tenant-wide ceilings — because each alone is bypassable. The
        tenant-wide ceiling is checked LAST, after parsing and validation, so
        malformed input cannot exhaust the day's allowance and deny the form to
        real people.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestBookingInput"
      responses:
        "202":
          description: >
            Accepted as a REQUEST. Returned identically whether or not the phone
            matches an existing patient — see the description. The body carries
            exactly one key and no match information of any kind.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [received]
        "400":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    patientBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        The patient's Supabase access token (JWT) carrying `role=patient`,
        `patient_id`, and `tenant_id` claims. The server derives identity from
        the verified token only; a staff token can never satisfy a patient route.
    ingestionHmac:
      type: apiKey
      in: header
      name: X-OsteoJP-Signature
      description: >
        Hex HMAC-SHA256 of `${timestamp}.${rawBody}` under the shared ingestion
        secret. MUST be sent together with `X-OsteoJP-Timestamp` (the
        `ingestionTimestamp` scheme) — the two form the HMAC pair.
    ingestionTimestamp:
      type: apiKey
      in: header
      name: X-OsteoJP-Timestamp
      description: Unix-seconds timestamp, bound into the signed string; must be within ±300s.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe's webhook signature over the raw body (provider-managed).
...
  responses:
...
    InvalidInput:
      description: "`invalid_input` — malformed/missing body fields."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorLocalized"
    RateLimited:
      description: >
        `rate_limited` — too many requests. Carries `Retry-After`. On the OTP
        paths the limiter is durable (Postgres) and FAILS CLOSED: if the store
        is unreachable this refuses rather than allowing, because failing open
        would turn a database blip into an unlimited guessing window against a
        6-digit code.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the window resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    GuestBookingInput:
      type: object
      additionalProperties: false
      required:
        [tenantId, fullName, phone, serviceId, locationId, preferredDate, preferredPeriod]
      description: >
        R-GUEST-2: the MINIMUM. A name, a number, the service, the clinic and
        WHEN THEY WOULD PREFER TO COME. NO NIF (PL-20), no health information,
        nothing clinical before the first visit.


        OPTION A, ratified 2026-08-14. This used to take `startsAt`/`endsAt` and
        an optional `practitionerId`, on the assumption the public form would
        offer real slots. It does not: the therapist roster and the slot list
        stay authenticated (exposure matrix MN-27, MN-28), so nothing public
        could have offered a caller a time or a therapist. An exact timestamp
        arriving here would therefore be UNSOURCED, and reception would read it
        as a time the person had been shown. Those three fields are now
        REFUSED — a body carrying them and nothing else is a 400.
      properties:
        tenantId:
          type: string
          format: uuid
        fullName:
          type: string
          maxLength: 120
          description: >
            Bounded so the reception queue cannot be filled with essays. The one
            free-text field on this surface.
        phone:
          type: string
          description: >
            Portuguese MOBILE number in any format; normalized to E.164 server
            side. A geographic (landline) number is refused: reception's only way
            to reach a guest is to ring them, so a landline typed by mistake is a
            request nobody can action.
          examples: ["+351912345678"]
        serviceId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        preferredDate:
          type: string
          format: date
          description: >
            The calendar day the caller would prefer, YYYY-MM-DD, read in
            Europe/Lisbon. Must not be in the past and must be within 90 days.
            A date that does not exist (2026-02-30) is refused rather than
            rolled forward.
          examples: ["2026-09-07"]
        preferredPeriod:
          type: string
          enum: [manha, tarde]
          description: >
            Morning (09:00–13:00) or afternoon (13:00–19:00), Europe/Lisbon,
            from the clinic's published opening hours.


            THIS IS A PREFERENCE AND NOT AN AVAILABILITY CLAIM. No schedule is
            consulted on this path and nothing is reserved. The pair
            (preferredDate, preferredPeriod) is encoded into the existing
            `requested_starts_at` / `requested_ends_at` columns by
            `encodeGuestPreferredWindow` in @osteojp/db — the window's
            boundaries, not a slot anybody was offered. Reception resolves the
            real time when they telephone.
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      description: Stable machine-readable error code (no localized message).
      properties:
        error:
          type: string
          examples: ["unauthorized"]
    ErrorLocalized:
      type: object
      additionalProperties: false
      required: [error, message]
      description: Stable error code plus a localized (PT-default) human message.
      properties:
        error:
          type: string
          examples: ["not_found"]
        message:
          type: string
...